Move Quartz Akka job code to common.
